Update dependencies, change API endpoint, and add state management in LoginScreen component
This commit updates the "axios" package to version 1.6.8, and adds several Expo modules such as "expo-camera", "expo-media-library", and "expo-notifications" to the project's dependencies. It also changes the API endpoint for the login route to a new one, and adds state management to the LoginScreen component using the "useState" hook. The commit also includes some changes to the UI components to reflect the state changes.

implementado o metodo de login e a conexão com a api